+++ src/index.ts
@@ -4,8 +4,8 @@ import HeapQueue from './HeapQueue';
 export const Int32Array_ID = idof<Int32Array>();
 
 
-const LARGE_VALUE = f32(1e6);
-const SMALL_VALUE = f32(1e-6);
+const LARGE_VALUE = f64(1e6);
+const SMALL_VALUE = f64(1e-6);
 const radius = i32(5);
 
 
@@ -20,7 +20,7 @@ function min<T>(vals: Array<T>): T {
 }
 
 
-function eikonal(n1: i32, n2: i32, flag: Int32Array, u: Float32Array): f32 {
+function eikonal(n1: i32, n2: i32, flag: Int32Array, u: Float64Array): f64 {
 	let u_out = LARGE_VALUE;
 	const u1 = unchecked(u[n1]);
 	const u2 = unchecked(u[n2]);
@@ -28,8 +28,8 @@ function eikonal(n1: i32, n2: i32, flag: Int32Array, u: Float32Array): f32 {
 	const f2 = unchecked(flag[n2]);
 	if (f1 == 0) {
 		if (f2 == 0) {
-			const perp = f32(Math.sqrt(2 - (u1 - u2) * (u1 - u2)));
-			let s = f32((u1 + u2 - perp) * 0.5);
+			const perp = f64(Math.sqrt(2 - (u1 - u2) * (u1 - u2)));
+			let s = f64((u1 + u2 - perp) * 0.5);
 			if (s >= u1 && s >= u2) {
 				u_out = s;
 			} else {
@@ -48,7 +48,7 @@ function eikonal(n1: i32, n2: i32, flag: Int32Array, u: Float32Array): f32 {
 }
 
 
-function grad_func(array: Float32Array, n: i32, step: i32, flag: Int32Array): f32 {
+function grad_func(array: Float64Array, n: i32, step: i32, flag: Int32Array): f64 {
 	if (unchecked(flag[n + step]) != 2) { // UNKNOWN
 		if (unchecked(flag[n - step]) != 2) {
 			return (unchecked(array[n + step] - array[n - step])) * 0.5;
@@ -68,14 +68,14 @@ function grad_func(array: Float32Array, n: i32, step: i32, flag: Int32Array): f3
 function inpaint_point(
 	n: i32,
 	flag: Int32Array,
-	u: Float32Array,
+	u: Float64Array,
 	image: Int32Array,
 	width: i32,
 	height: i32,
 	indices_centered: Array<i32>
 ): void {
-	let Ia: f32 = 0;
-	let norm: f32 = 0;
+	let Ia: f64 = 0;
+	let norm: f64 = 0;
 
 	const gradx_u = grad_func(u, n, 1, flag);
 	const grady_u = grad_func(u, n, width, flag);
@@ -101,17 +101,17 @@ function inpaint_point(
 			continue;
 		}
 
-		const rx = f32(i - i_nb);
-		const ry = f32(j - j_nb);
+		const rx = f64(i - i_nb);
+		const ry = f64(j - j_nb);
 		const rx2 = rx * rx;
 		const ry2 = ry * ry;
 
-		const geometric_dst = f32(1 / ((rx2 + ry2) * Math.sqrt(rx2 + ry2)));
-		const levelset_dst = f32(1 / (1 + Math.abs(unchecked(u[nb]) - unchecked(u[n]))));
-		const direction = f32(Math.abs(rx * gradx_u + ry * grady_u));
-		const weight: f32 = geometric_dst * levelset_dst * direction + SMALL_VALUE;
+		const geometric_dst = f64(1 / ((rx2 + ry2) * Math.sqrt(rx2 + ry2)));
+		const levelset_dst = f64(1 / (1 + Math.abs(unchecked(u[nb]) - unchecked(u[n]))));
+		const direction = f64(Math.abs(rx * gradx_u + ry * grady_u));
+		const weight: f64 = geometric_dst * levelset_dst * direction + SMALL_VALUE;
 
-		Ia += weight * f32(unchecked(image[nb]));
+		Ia += weight * f64(unchecked(image[nb]));
 		norm += weight;
 	}
 	// the fmm.py which this is based on actually implements a slightly different
@@ -135,7 +135,7 @@ function inpaintChannel(
 ): void {
 	const size = width * height;
 	const flag = new Int32Array(size);
-	const u = new Float32Array(size);
+	const u = new Float64Array(size);
 
 	for (let i = 0; i < size; i++) {
 		if (!unchecked(mask[i])) {
